refactor: geo-coding 서비스에서 geo-coding 요청 클라이언트를 분리
* geo-coding API를 호출하는 클라이언트를 서비스에서 분리
  - 테스트 코드에서 mocking 대상을 명시적으로 제시함.
  - 네이버 클라우드 외의 geo-coding API를 염두에 둠.
    - 단, 현재 geo-coding 서비스의 로직이 네이버 클라우드에 종속된
      상태임.
