+++ api/src/main/java/burgermap/service/geoCoding/GeoCodingClient.java
@@ -1,0 +1,5 @@
+package burgermap.service.geoCoding;
+
+public interface GeoCodingClient {
+    public String request(String address);
+}
